+++ lib/dartarabic.dart
@@ -1,5 +1,5 @@
 library dartarabic;
 
 export 'src/dartarabic.dart';
-export 'src/arabic.dart';
+export 'src/arabic/_lib.dart';
 export 'src/x3orouby.dart';